Refactor packaging featurewise
