+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.MicroFeatureCodelab"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.MainActivity"
+            android:name=".personalisedjob.presentation.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.MicroFeatureCodelab">
