+++ app/src/main/java/com/example/microfeaturecodelab/base/ComponentViewModelFactory.kt
@@ -1,35 +1,26 @@
 package com.example.microfeaturecodelab.base
 
 import android.util.Log
-import com.example.microfeaturecodelab.presentation.featureconfig.ComponentConfig
-import com.example.microfeaturecodelab.presentation.uimodel.PersonalisedJobViewModel
+import com.example.microfeaturecodelab.personalisedjob.presentation.featureconfig.ComponentConfig
 import kotlinx.coroutines.CoroutineScope
 import javax.inject.Inject
 
 class ComponentViewModelFactory @Inject constructor(
     private val viewModelFactories: Map<String, @JvmSuppressWildcards MicroFeatureFactory<out MicroFeatureViewModel>>,
     private val composerFactories: Map<String, @JvmSuppressWildcards AbstractMicroFeatureComposer<out MicroFeatureViewModel>>,
-    private val featureFactory: PersonalisedJobViewModel.Factory
 ) {
 
     fun create(
         componentConfig: ComponentConfig,
         viewModelScope: CoroutineScope
-    ): MicroFeatureViewModel {
-        return when (componentConfig.type) {
-            "personalisedJob" -> {
-                Log.d("ComponentViewModelFactory", "VM: ${viewModelFactories[componentConfig.type]}")
-                Log.d("ComponentViewModelFactory", "Composer: ${composerFactories[componentConfig.type]}")
-                val viewmodel = featureFactory.create(viewModelScope)
-                viewmodel
-            }
-            "premiumApplicant" -> {
-                Log.d("ComponentViewModelFactory", "Composer: ${composerFactories[componentConfig.type]}")
-                val viewmodel = featureFactory.create(viewModelScope)
-                viewmodel
-            }
-            else -> throw IllegalArgumentException("Unknown component type: ${componentConfig.type}")
+    ): Pair<MicroFeatureViewModel?, AbstractMicroFeatureComposer<out MicroFeatureViewModel>?> {
+        return viewModelFactories[componentConfig.type]?.let {
+            Log.d("ComponentViewModelFactory", "Creating viewmodel for ${componentConfig.type}")
+            val componentVM = viewModelFactories[componentConfig.type]?.create(viewModelScope)
+            val componentComposer = composerFactories[componentConfig.type]
+            componentVM to componentComposer
+        } ?: kotlin.run {
+            throw IllegalArgumentException("Unknown component type: ${componentConfig.type}")
         }
     }
-
 }
