+++ app/src/main/java/com/example/microfeaturecodelab/di/ComponentViewModelModule.kt
@@ -2,7 +2,8 @@ package com.example.microfeaturecodelab.di
 
 import com.example.microfeaturecodelab.base.MicroFeatureFactory
 import com.example.microfeaturecodelab.base.MicroFeatureViewModel
-import com.example.microfeaturecodelab.presentation.uimodel.PersonalisedJobViewModel
+import com.example.microfeaturecodelab.premiumApplicant.PremiumApplicantViewModel
+import com.example.microfeaturecodelab.personalisedjob.presentation.uimodel.PersonalisedJobViewModel
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -17,5 +18,10 @@ interface ComponentViewModelModule {
     @IntoMap
     @ComponentTypeKey("personalisedJob")
     fun bindPersonalisedViewModel(viewmodelFactory: PersonalisedJobViewModel.Factory): MicroFeatureFactory<out MicroFeatureViewModel>
+
+    @Binds
+    @IntoMap
+    @ComponentTypeKey("premiumApplicant")
+    fun bindPremiumApplicantViewModel(viewmodelFactory: PremiumApplicantViewModel.Factory): MicroFeatureFactory<out MicroFeatureViewModel>
 
 }
