+++ app/src/main/java/com/example/microfeaturecodelab/personalisedjob/domain/JobsRepository.kt
@@ -1,0 +1,10 @@
+package com.example.microfeaturecodelab.personalisedjob.domain
+
+import com.example.microfeaturecodelab.personalisedjob.domain.model.JobRecommendations
+import kotlinx.coroutines.flow.Flow
+
+interface JobsRepository {
+
+    fun fetchJobs(input: JobQueryParameter): Flow<JobRecommendations>
+}
+
