+++ app/src/main/java/com/example/microfeaturecodelab/personalisedjob/presentation/model/JobItem.kt
@@ -1,8 +1,8 @@
-package com.example.microfeaturecodelab.presentation.model
+package com.example.microfeaturecodelab.personalisedjob.presentation.model
 
 import androidx.compose.runtime.Stable
-import com.example.microfeaturecodelab.domain.model.JobRecommendation
-import com.example.microfeaturecodelab.domain.model.JobRecommendations
+import com.example.microfeaturecodelab.personalisedjob.domain.model.JobRecommendation
+import com.example.microfeaturecodelab.personalisedjob.domain.model.JobRecommendations
 
 @Stable
 data class RecommendedJobSection(
