+++ app/src/main/java/com/example/microfeaturecodelab/premiumApplicant/PremiumApplicantComposer.kt
@@ -3,12 +3,14 @@ package com.example.microfeaturecodelab.premiumApplicant
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.example.microfeaturecodelab.base.AbstractMicroFeatureComposer
+import com.example.microfeaturecodelab.personalisedjob.presentation.featureconfig.ComponentConfig
 import javax.inject.Inject
 
 class PremiumApplicantComposer @Inject constructor(): AbstractMicroFeatureComposer<PremiumApplicantViewModel>() {
     @Composable
     override fun Content(
         viewModel: PremiumApplicantViewModel,
+        config: ComponentConfig,
         modifier: Modifier,
         onAction: (Any) -> Unit
     ) {
